requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
PyMySQL==1.1.0
transformers==4.21.3
deep-translator==1.11.4
gtts==2.3.2
moviepy==1.0.3
webvtt-py==0.4.6